pull crypto wallet into ints own file component and display basic portfolio values
