--- client/src/styles/_home/_profileDrawer.scss
@@ -1,5 +1,0 @@
-.ProfileDrawer {
-}
-
-.ProfileBox {
-}
